+++ main/src/main/scala/org/clulab/reach/brat/Brat.scala
@@ -13,12 +13,16 @@ import org.clulab.reach.utils.BratUtils
 import org.clulab.struct.Interval
 
 object Brat extends LazyLogging {
+  /* // old version of the readStandoff method
   def readStandOff(input: String): Seq[Annotation] = {
     input
     .split("\n")
     .toSeq
     .flatMap{ parseAnnotation(_) }
   }
+  */
+  def readStandOff(input: String): Seq[Annotation] =
+    input.split('\n').map(_.trim).flatMap(parseAnnotation) // getLines doesn't work in Java 11.
 
   def readStandOff(input: InputStream): Seq[Annotation] =
     Source.fromInputStream(input).getLines.toSeq flatMap parseAnnotation